backend/requirements.txt
aiohappyeyeballs==2.3.2
aiohttp==3.10.0
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.37.1
anyio==4.4.0
appdirs==1.4.4
appnope==0.1.4
archspec @ file:///home/conda/feedstock_root/build_artifacts/archspec_1708969572489/work
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.13.3
bleach==6.2.0
blessed==1.20.0
blinker==1.8.2
boltons @ file:///home/conda/feedstock_root/build_artifacts/boltons_1711936407380/work
Brotli @ file:///Users/runner/miniforge3/conda-bld/brotli-split_1725267563793/work
CacheControl==0.14.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
comm==0.2.2
conda @ file:///Users/runner/miniforge3/conda-bld/conda_1727791778282/work
conda-libmamba-solver @ file:///home/conda/feedstock_root/build_artifacts/conda-libmamba-solver_1727359833193/work/src
conda-package-handling @ file:///home/conda/feedstock_root/build_artifacts/conda-package-handling_1729006966809/work
conda_package_streaming @ file:///home/conda/feedstock_root/build_artifacts/conda-package-streaming_1729004031731/work
contourpy==1.3.0
cryptography==43.0.0
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
distro @ file:///home/conda/feedstock_root/build_artifacts/distro_1704321475663/work
editor==1.6.6
et-xmlfile==1.1.0
executing==2.1.0
fastapi==0.95.2
fastjsonschema==2.21.1
filelock==3.16.1
fonttools==4.54.1
fqdn==1.5.1
frozendict @ file:///Users/runner/miniforge3/conda-bld/frozendict_1728841362247/work
frozenlist==1.4.1
fsspec==2024.10.0
git-python==1.0.3
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-firestore==2.17.0
google-cloud-storage==2.18.0
google-crc32c==1.5.0
google-generativeai==0.7.2
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.67.0
grpcio-status==1.62.2
h11==0.14.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1634280454336/work
hpack==4.0.0
html2image==2.0.5
html2text==2024.2.26
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.26.2
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1619110129307/work
idna==3.7
importlib_metadata==8.5.0
inquirer==3.4.0
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
json5==0.10.0
jsonpatch @ file:///home/conda/feedstock_root/build_artifacts/jsonpatch_1695536281965/work
jsonpointer @ file:///Users/runner/miniforge3/conda-bld/jsonpointer_1725302931775/work
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
langchain==0.2.11
langchain-community==0.2.10
langchain-core==0.2.41
langchain-experimental==0.0.63
langchain-google-genai==1.0.8
langchain-openai==0.1.19
langchain-text-splitters==0.2.4
langsmith==0.1.137
libmambapy @ file:///Users/runner/miniforge3/conda-bld/mamba-split_1725066179274/work/libmambapy
litellm==1.40.8
llvmlite==0.43.0
mamba @ file:///Users/runner/miniforge3/conda-bld/mamba-split_1725066179274/work/mamba
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
menuinst @ file:///Users/runner/miniforge3/conda-bld/menuinst_1725359012839/work
mistune==3.1.1
monotonic==1.6
more-itertools==10.5.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.3.2
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
open-interpreter==0.2.0
openai==1.52.2
openai-whisper @ git+https://github.com/openai/whisper.git@90db0de1896c23cbfaf0c58bc2d30665f709f170
openpyxl==3.1.5
orjson==3.10.6
outcome==1.3.0.post0
overrides==7.7.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1718189413536/work
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1726613481435/work
pluggy @ file:///home/conda/feedstock_root/build_artifacts/pluggy_1713667077545/work
posthog==3.7.4
prometheus_client==0.21.1
prompt-toolkit==3.0.43
proto-plus==1.24.0
protobuf==4.25.4
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAutoGUI==0.9.54
pycosat @ file:///Users/runner/miniforge3/conda-bld/pycosat_1696355905762/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1711811537435/work
pydantic==2.7.4
pydantic_core==2.18.4
PyGetWindow==0.0.9
Pygments==2.18.0
PyJWT==2.8.0
PyMsgBox==1.0.9
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-Quartz==10.3.1
pyparsing==3.1.2
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
pytweening==1.2.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.2.0
readchar==4.2.0
referencing==0.35.1
regex==2024.7.24
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1717057054362/work
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.3
rpds-py==0.20.0
rsa==4.9
ruamel.yaml @ file:///Users/runner/miniforge3/conda-bld/ruamel.yaml_1728764996801/work
ruamel.yaml.clib @ file:///Users/runner/miniforge3/conda-bld/ruamel.yaml.clib_1728724504988/work
rubicon-objc==0.4.9
runs==1.2.2
selenium==4.25.0
Send2Trash==1.8.3
setuptools==75.1.0
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.31
stack-data==0.6.3
starlette==0.27.0
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.3.0
terminado==0.18.1
tiktoken==0.4.0
tinycss2==1.4.0
tokenizers==0.20.1
tokentrim==0.1.13
toml==0.10.2
torch==2.5.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
truststore @ file:///home/conda/feedstock_root/build_artifacts/truststore_1724770958874/work
typer==0.12.5
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.34.0
wcwidth==0.2.13
webcolors==24.11.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wget==3.2
wheel==0.44.0
widgetsnbextension==4.0.13
wsproto==1.2.0
xmod==1.8.1
yarl==1.9.4
yaspin==3.1.0
zipp==3.20.2
zstandard==0.23.0
